+++ pkg/apis/infrastructure/v1alpha2/network_pool_funcs.go
@@ -1,0 +1,24 @@
+package v1alpha2
+
+import (
+	v1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+
+	"github.com/giantswarm/apiextensions/v2/pkg/crd"
+)
+
+const (
+	kindNetworkPool              = "NetworkPool"
+	networkpoolDocumentationLink = "https://docs.giantswarm.io/reference/cp-k8s-api/networkpools.infrastructure.giantswarm.io/"
+)
+
+func NewNetworkPoolCRD() *v1.CustomResourceDefinition {
+	return crd.LoadV1(group, kindNetworkPool)
+}
+
+func NewNetworkPoolTypeMeta() metav1.TypeMeta {
+	return metav1.TypeMeta{
+		APIVersion: SchemeGroupVersion.String(),
+		Kind:       kindNetworkPool,
+	}
+}
